feat(ringo-flow): QuickJS (JS/TS) scenario frontend
A JavaScript/TypeScript scenario frontend over the shared engine, alongside
rhai. Selected by a .js file (or --lang js); both frontends implement the same
ScriptHost.

DSL parity with rhai: agent verbs/getters, dial/transfer/deflect overloads
(Agent | string), Jest-style assertions (expect().toBe/.not/...), http client,
mock server (static + dynamic responders), audio send/verify, scenario/setup/
teardown, tags/skip/only, --set/--env-file layering + loadEnv, ES-module import,
and free-form agent metadata. Blocking waiters (until/verifyAudio/http/wait) are
native Promises, so Promise.all runs them concurrently.

Runs on an rquickjs AsyncRuntime/AsyncContext driven by futures-executor on the
scenario thread.

Editor support: the .d.ts is generated from the Rust binding signatures at
compile time via ringo-flow-macros (#[ts_export]/#[ts_global]/#[derive(TsInterface
|TsEnum)] + declare!), collected through inventory and rendered by tsgen. One
canonical copy in docs/, guarded by a drift test and a tsc CI gate that
type-checks the example scenarios.

+++ crates/ringo-flow/src/script/rhai/types.rs
@@ -345,16 +345,16 @@ impl Assertion {
         self.chain(self.inner.matches(pattern))
     }
     pub(super) fn greater_than(&mut self, n: i64) -> Check {
-        self.chain(self.inner.greater_than(n))
+        self.chain(self.inner.greater_than(n as f64))
     }
     pub(super) fn at_least(&mut self, n: i64) -> Check {
-        self.chain(self.inner.at_least(n))
+        self.chain(self.inner.at_least(n as f64))
     }
     pub(super) fn less_than(&mut self, n: i64) -> Check {
-        self.chain(self.inner.less_than(n))
+        self.chain(self.inner.less_than(n as f64))
     }
     pub(super) fn at_most(&mut self, n: i64) -> Check {
-        self.chain(self.inner.at_most(n))
+        self.chain(self.inner.at_most(n as f64))
     }
 }
